Hello world
This uses the Client class to connect to Serial and send GCODE commands. The HandDetector class uses cv2 to capture video and mediapipe to detect hands on screen. The main script interprets the hand movement to send GCODE commands based on the position of the hand on screen.
Not implemented: Z axis to determine pen up or down
Issues: video freezes and stutters while gcode is being sent